+++ packages/web3-eth/CHANGELOG.md
@@ -201,6 +201,7 @@ Documentation:
 
 -   Ensure provider.supportsSubscriptions exists before watching by subscription (#6440)
 -   Fixed param sent to `checkRevertBeforeSending` in `sendSignedTransaction` 
+-   Fixed `defaultTransactionBuilder` for value issue (#6509)
 
 ### Added